requirements.txt
discord.py
tortoise-orm
asyncpg